+++ CHANGELOG.md
@@ -10,6 +10,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Trigger events in value extractors to allow other modules to add fields
+- Added ability to declare search fields
+- Added ability to define `qf` and `mm` parameters
+- Implemented several search operators (contains any word, contains all words,
+  is like)
 
 
 ## [0.5.0] - 2019-03-06